+++ src/effects/spells/greenFlameBlade.js
@@ -6,6 +6,7 @@ export async function greenFlameBladeEffect(document) {
   document.system.range = { value: null, units: "self", long: null };
   document.system.actionType = "other";
   document.system.save.ability = "";
+  document.system.chatFlavor = "";
   DDBMacros.setMidiOnUseMacroFlag(document, "spell", "greenFlameBlade.js", ["postActiveEffects"]);
   await DDBMacros.setItemMacroFlag(document, "spell", "greenFlameBlade.js");
   return document;